Update error handling for the lidar homing updates.
Now, if we fail to do initialize landmark when we start, we instead initialize the landmark on the first scheduled homing update.

within the lidar homing update, we first try multiple look angles. If these fail, we try to process the unfiltered lidar. WHen using the unfiltered lidar, there are cases where the solution is bad.  when solution is bad, so we monitor this, and just get the base location on the next homing update.  the first homing update typically has meter-level error, but his correction is good enough to keep the run alive and better than just having bogus updates.

we could try recovery manuevers during init, but this wastes time. It is rare enough, that I am ok, with just not init landmark and moving on.

now, when a normal homing update fails, we just keep going.

was lucky enough to find a seed to test all of this logic 1599760402